+++ src/Accel_Vector.cpp
@@ -7,7 +7,7 @@
 
 #include "../headers/Accel_Vector.h"
 
-Accel_vector::Accel_vector(int32_t Accel_x,int32_t Accel_y,int32_t Accel_z,int32_t Accel_a,uint64_t epoch_start,uint64_t epoch_stop)//,uint32_t epoch_stop
+Accel_vector::Accel_vector(int16_t Accel_x,int16_t Accel_y,int16_t Accel_z,int16_t Accel_a,uint64_t epoch_start,uint64_t epoch_stop)//,uint32_t epoch_stop
 			:
 		Accel_x(Accel_x), Accel_y(Accel_y), Accel_z(Accel_z), Accel_a(Accel_a), epoch_start(
 				epoch_start),epoch_stop(epoch_stop) { //, epoch_stop(epoch_stop)
